notes/requirements.txt
python-dotenv
cryptography
PyMySQL